refactor: :recycle: Découpe l’app demo en plus de pages
